requirements.txt
torch
torchvision
transformers
pillow
gradio
tqdm
opencv-python
numpy
